add homework manager

+++ src/main/java/me/Mung/Commands/LostArk/CommandRm.java
@@ -2,8 +2,10 @@ package me.Mung.Commands.LostArk;
 
 import me.Mung.Model.PlayerDAO;
 import me.Mung.Model.PlayerVO;
+import me.Mung.type.Command;
 import me.Mung.type.SlashCommand;
 import net.dv8tion.jda.api.entities.Member;
+import net.dv8tion.jda.api.entities.Message;
 import net.dv8tion.jda.api.entities.TextChannel;
 import net.dv8tion.jda.api.events.interaction.ButtonClickEvent;
 import net.dv8tion.jda.api.events.interaction.SelectionMenuEvent;
@@ -12,32 +14,22 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.util.Objects;
-import java.util.concurrent.TimeUnit;
 
-public class SlashCommandRm implements SlashCommand {
-    private static final Logger LOGGER = LoggerFactory.getLogger(SlashCommandRm.class);
+public class CommandRm implements Command {
+    private static final Logger LOGGER = LoggerFactory.getLogger(CommandRm.class);
 
     //  해당 캐릭터를 DB에 삭제시킨다.
     @Override
-    public void performCommand(SlashCommandEvent event, Member m, TextChannel channel) {
+    public void performCommand(Member m, TextChannel channel, Message message) {
         LOGGER.info(m.getUser().getAsTag());
-
+        String[] command = message.getContentRaw().split(" ");
+        if (command.length != 2)
+            return;
         PlayerVO user = new PlayerVO();
         user.setId_name(m.getId());
-        user.setChar_name(Objects.requireNonNull(event.getOption("character")).getAsString());
+        user.setChar_name(command[1]);
         user.setCur_level(1.);
         PlayerDAO.deletePlayer(user);
-        LOGGER.info("{}", user);
-        event.reply(user.getChar_name() + " 삭제").setEphemeral(true).queue();
-    }
-
-    @Override
-    public void performCommand(ButtonClickEvent event, Member m, TextChannel channel) {
-
-    }
-
-    @Override
-    public void performCommand(SelectionMenuEvent event, Member m, TextChannel channel) {
-
+        message.reply(m.getAsMention() + " " + user.getChar_name() + " 삭제").queue();
     }
 }
